+++ cli/disk_cache.rs
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
-use crate::fs as deno_fs;
+use crate::fs_util;
 use crate::http_cache::url_to_filename;
 use deno_core::url::{Host, Url};
 use std::ffi::OsStr;
@@ -145,7 +145,7 @@ impl DiskCache {
       Some(ref parent) => self.ensure_dir_exists(parent),
       None => Ok(()),
     }?;
-    deno_fs::write_file(&path, data, crate::http_cache::CACHE_PERM)
+    fs_util::write_file(&path, data, crate::http_cache::CACHE_PERM)
       .map_err(|e| with_io_context(&e, format!("{:#?}", &path)))
   }
 }
