+++ cli/test_runner.rs
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
-use crate::fs as deno_fs;
+use crate::fs_util;
 use crate::installer::is_remote_url;
 use deno_core::error::AnyError;
 use deno_core::serde_json::json;
@@ -42,10 +42,10 @@ pub fn prepare_test_modules_urls(
   let mut prepared = vec![];
 
   for path in include_paths {
-    let p = deno_fs::normalize_path(&root_path.join(path));
+    let p = fs_util::normalize_path(&root_path.join(path));
     if p.is_dir() {
       let test_files =
-        crate::fs::collect_files(vec![p], vec![], is_supported).unwrap();
+        crate::fs_util::collect_files(vec![p], vec![], is_supported).unwrap();
       let test_files_as_urls = test_files
         .iter()
         .map(|f| Url::from_file_path(f).unwrap())
